requirements.txt
jsonschema==4.6.1
ruamel.yaml==0.17.21
requests==2.31.0